Add override to prevent kube-proxy adjusting conntrack settings
In a containerised environment (docker/LXC/LXD...) with a non
net-init namespace it is not possible to adjust conntrack settings.
kube-proxy attempts to do this and fails to start unless it is
configured not to do so.

The previous logic which detected the use of a docker type ansible
connection plugin is converted to an overridable variable to allow
a deployment tool to specify that conntrack should not be adjusted.

See https://github.com/kubernetes/kubernetes/pull/103174