Define default version for kubeadm/kubectl/kubelet
The version vars for these tools are used in the defaults file
of their respective role, but is not defined, normally overidden
to the value of 'kubernetes_version' when the role is called.

This patch sets the default value of the installed version to
be the newest one understood by the defaults file.