fix: Check that resources are in kubeadm-config configmap before using them
In some conditions the k8s_info module can succeed but return and
empty list of resources. This can happen when the api server is
running but the kubeadm-config cm is missing, then the next tasks
fail with undefined variables.

This patch extends the condition on the kubernetes version check to
ensure that resources from kubeadm-config are present.