Allow pip packages to be optionally installed into a virtualenv
pep-668 will force installation of python modules with pip into
a virtualenv rather than the system python. This is already required
for debian-12 and ubuntu-24.04.

This patch allows the caller of ansible-collection-kubernetes to
pass the path to a pre-created virtualenv to install python target
host requirements for ansible modules into.